feat(presence-update): add presence-update task
Add a new task to update the bot's presence every 10 minutes. The presence
will cycle through a list of predefined activities, including watching the
number of servers, available users, unique users, and the world, and listening
to the questions. The task will also replace the {UNIQUE_USERS} placeholder in
the activity message with the number of unique users in the database.

feat(presence-update): add Activity type

Add a new type to represent the bot's activity, which includes the activity
type and message.

feat(function.ts): add getAvaibleUsers and getUniqueUsers functions

Add two new functions to get the number of available users and unique users in
the database, respectively.

refactor(task.type.ts): add MaybePromise type

Add a new type to represent a value that may or may not be a promise.

refactor(task/index.ts): export TaskInterval and Task