feat(guild-member-add.event.ts): add guild member add event handler
This commit adds a new event handler for the guild member add event. When a new member joins a guild, the bot will send a message to a specific support channel in a support guild. The message contains the name and ID of the guild the bot joined, as well as the current number of guilds the bot is in.