feat(guild-member-delete.event.ts): add guild member delete event handler
This commit adds a new event handler for the "guildDelete" event. When a guild is deleted, the bot will send a message to a specific support channel in another guild, indicating that the guild has been deleted. The message includes the name and ID of the deleted guild, as well as the current number of guilds the bot is in.