feat(privacy): add delete data feature
This commit adds a new feature to the privacy command that allows users to delete all their data from the bot. The feature is implemented with a button that, when clicked, prompts the user to confirm the action. If the user confirms, all their data is deleted from the database. The commit also includes a new function to delay the execution of the code.