feat(learn): add learn command
Add learn command with support for multiple languages and a slash command builder. The command sends a private message to the user every day at 12:00 (UTC+2) with information about the world around them. The user can react to the message with a thumbs up or thumbs down to indicate whether they liked the information or not. The command also includes support for translations and images.